Move instruction cache to the processor and rename it execution cache.
The previous instruction cache implementation was caching virtual
address translations and decoded instructions in the mmu, and it
included a separate function call in the step function for each entry in
the mmu's cache.

This commit moves the mmu portion of the cache implementation to the
processor_t class to clean up the interactions between the processor and
the mmu.

As a result of this, the following has changed:
* decoding is now entirely done in the processor
* the flush_icache method in mmu_t doesn't do anything
* insn_fetch_t no longer includes inst_func_t, but does include if the
  result can be cached
* added a fence_i method for processor_t and changed the implementation
  of fence.i to call that method instead